+++ spec/opal/core/string_spec.rb
@@ -1,0 +1,11 @@
+require 'spec_helper'
+
+describe "String" do
+  it "handles contiguous parts correctly" do
+    str = "a" "b"
+    str.should == "ab"
+
+    str2 = "d" "#{str}"
+    str2.should == "dab"
+  end
+end