+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 *   Fix Range.new to raise `ArgumentError` on contructor values that cannot
     be compared
 
+*   Fix compiler bug where Contiguous strings were not getting concatenated.
+
 ## 0.6.2 2014-04-25
 
 *   Added Range#size